* Implemented request "1662062 - Hyperlink.openLocation: Remove Toolbars". Added support for this feature at multiple levels. Hyperlink.openLocation now has two additional overloads to support the new 'visibleChrome' and 'resizeAllowed' options.  By specifying 'false' for visibleChrome, the resulting popup window will have it's menubar, status, toolbar, locationbar and directoriesbar stripped provided the browser supports it.  Likewise, specifying false for resizeAllowed will prevent the popup window from being resized.
Additionally, three new properties have been added to the Hyperlink class to enable the Hyperlink component to mimic the same feature set that openLocation offers.  Hyperlink has been missing the ability to control the target window for sometime, so that is now supported via the 'target' property.  Further, there are now 'visibleChrome' and 'resizeAllowed' properties that cause similar behavior to the new openLocation features.

* Added extra 'null' check in WebServlet to prevent a NullPointerException that sometimes occurs when refreshing while the app server is performing a context reload.