* Implemented feature request "1841802 - Support shared use of the thinwire.jar, extend WebServlet".  You can now subclass WebServlet to create a servlet entry point within your application.  For example, if your application package is com.mycompany.myapp, you might create a class called AppContextServlet that subclasses WebSevlet. Then you would configure your application's web.xml so that the servlet is AppContextServlet. If you do something similar for all of your ThinWire applications, you should then be able to have a single thinwire.jar that is placed in the shared library folder of the application server, which is 'shared/lib' in the case of Tomcat.
As an additional bonus, if you exclude the 'mainClass' init-param and add a 'main' method to the subclassed AppContextServlet, the framework will now treat that as the entry point for your ThinWire application.  This makes it a little easier to setup the bootstrap code for your application.  Keep in mind that you CANNOT construct an instance of your subclassed AppContextServlet, servlets are managed by the application server only.  So, only use the 'main' entry point on AppContextServlet to construct other object instances and to configure the initial basic state of your application.  Because of this issue and the likelihood that someone would do so inadvertently, ThinWire has been setup to throw an exception if you create an instance of the subclassed servlet in your application code.