Removed the static Dialog methods of FileChooser
I think this was one of those features that was driven by a CCS specific
need. If someone wants to put a FileChooser in a Dialog there is nothing
stoppint them, so I don't feel bad about removing this.

Mainly, this depended on blocking Dialogs, and rather than re-write it,
it seemed like a good time to just get rid of the feature.