MessageBox.confirm no longer blocks
This was a tricky one. The solution for now is to have the confirm
methods take an ActionListener as a parameter. When the MessageBox is
closed, the ActionListener executes the actionPerformed() method. The
ActionEvent's source property is set to the id of the Button that was
clicked (this used to be the return value of confirm())

I'm not sold on this approach. Part of me wanted to just get rid of it,
but too much testing code is dependent on blocking MessageBoxes right
now. MessageBoxes with callbacks takes care of my test cases, so that's
what I'm going with for the moment.