Bumping the version number to 1.3-SNAPSHOT
Yes, I am *very* aware that there never was a 1.2 final release.
However, I'm feeling the desire to start hacking around on the framework
again, and I want the freedom to change APIs and behaviors. We had
frozen features and APIs for 1.2 back in 2007, so it just seems
responsible to bump the version number before I start screwing around
with things.