Rewrote BorderImage to use CSS border-images
This component was written to mirror the behavior of the draft (at the
time) spec for css border-images, so there are no api changes. There's
probably still a bunch of unnecessary boilerplate code that can be
removed, but this was a fairly unobtrusive way to implement this change.

Also, I'm seeing what it's like to use Coffeescript instead of
Javascript for writing the client-side components. So far, I like it.